feat: Show player name when giving all pieces in game
The code changes in `game.js` add an event listener for the "game:givePieces" event. When this event is triggered, the player's name is retrieved based on their ID and displayed in the game UI with the message "{playerName} Gave All Pieces". This provides a clear indication of which player gave all their pieces.